Move Framework\Helpers namespace to \Hyde\Helpers
