+++ packages/framework/config/hyde.php
@@ -12,8 +12,8 @@
 */
 
 use Hyde\Blogging\Models\Author;
-use Hyde\Framework\Helpers\Features;
-use Hyde\Framework\Helpers\Meta;
+use Hyde\Helpers\Features;
+use Hyde\Helpers\Meta;
 
 return [
     /*
