+++ packages/framework/resources/views/components/post/article.blade.php
@@ -1,15 +1,16 @@
-<article aria-label="Article" id="{{ Hyde::url("posts/$page->identifier", '') }}" itemscope itemtype="http://schema.org/Article"
-    @class(['post-article mx-auto prose dark:prose-invert', 'torchlight-enabled' => Hyde\Framework\Helpers\Features::hasTorchlight()])>
+<article aria-label="Article" id="{{ Hyde::url("posts/$page->identifier", '') }}" itemscope
+         itemtype="http://schema.org/Article"
+        @class(['post-article mx-auto prose dark:prose-invert', 'torchlight-enabled' => Hyde\Helpers\Features::hasTorchlight()])>
     <meta itemprop="identifier" content="{{ $page->identifier }}">
     @if(Hyde::hasSiteUrl())
-    <meta itemprop="url" content="{{ Hyde::url('posts/' . $page->identifier) }}">
+        <meta itemprop="url" content="{{ Hyde::url('posts/' . $page->identifier) }}">
     @endif
-    
+
     <header aria-label="Header section" role="doc-pageheader">
         <h1 itemprop="headline" class="mb-4">{{ $page->title }}</h1>
-		<div id="byline" aria-label="About the post" role="doc-introduction">
+        <div id="byline" aria-label="About the post" role="doc-introduction">
             @includeWhen(isset($page->date), 'hyde::components.post.date')
-		    @includeWhen(isset($page->author), 'hyde::components.post.author')
+            @includeWhen(isset($page->author), 'hyde::components.post.author')
             @includeWhen(isset($page->category), 'hyde::components.post.category')
         </div>
     </header>
