+++ packages/framework/src/Framework/Services/MarkdownService.php
@@ -8,7 +8,7 @@ use Hyde\Framework\Actions\MarkdownConverter;
 use Hyde\Framework\Concerns\Internal\SetsUpMarkdownConverter;
 use Hyde\Framework\Contracts\MarkdownPostProcessorContract as PostProcessor;
 use Hyde\Framework\Contracts\MarkdownPreProcessorContract as PreProcessor;
-use Hyde\Framework\Helpers\Features;
+use Hyde\Helpers\Features;
 use Hyde\Pages\DocumentationPage;
 use League\CommonMark\Extension\DisallowedRawHtml\DisallowedRawHtmlExtension;
 use League\CommonMark\Extension\HeadingPermalink\HeadingPermalinkExtension;
