+++ packages/framework/src/Framework/Services/SemanticDocumentationArticle.php
@@ -4,7 +4,7 @@ declare(strict_types=1);
 
 namespace Hyde\Framework\Services;
 
-use Hyde\Framework\Helpers\Features;
+use Hyde\Helpers\Features;
 use Hyde\Pages\DocumentationPage;
 use Illuminate\Support\Str;
