+++ packages/framework/src/Helpers/Asset.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace Hyde\Framework\Helpers;
+namespace Hyde\Helpers;
 
 use Hyde\Framework\Services\AssetService;
 use Illuminate\Support\Facades\Facade;
