+++ packages/framework/src/Helpers/Features.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace Hyde\Framework\Helpers;
+namespace Hyde\Helpers;
 
 use Hyde\Framework\Concerns\JsonSerializesArrayable;
 use Hyde\Framework\Services\DiscoveryService;
