+++ packages/framework/src/Helpers/Includes.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace Hyde\Framework\Helpers;
+namespace Hyde\Helpers;
 
 use Hyde\Hyde;
 use Hyde\Markdown\Models\Markdown;
