+++ packages/framework/src/Hyde.php
@@ -9,7 +9,7 @@ use Hyde\Foundation\HydeKernel;
 use Hyde\Foundation\PageCollection;
 use Hyde\Foundation\RouteCollection;
 use Hyde\Framework\Concerns\HydePage;
-use Hyde\Framework\Helpers\Features;
+use Hyde\Helpers\Features;
 use Hyde\Support\Route;
 use Illuminate\Support\Facades\Facade;
