+++ packages/framework/src/Metadata/GlobalMetadataBag.php
@@ -5,9 +5,9 @@ declare(strict_types=1);
 namespace Hyde\Metadata;
 
 use Hyde\Framework\Concerns\HydePage;
-use Hyde\Framework\Helpers\Features;
-use Hyde\Framework\Helpers\Meta;
 use Hyde\Framework\Services\RssFeedService;
+use Hyde\Helpers\Features;
+use Hyde\Helpers\Meta;
 use Hyde\Hyde;
 use Illuminate\Support\Facades\View;
