+++ packages/framework/tests/Unit/Views/StylesComponentViewTest.php
@@ -4,7 +4,7 @@ declare(strict_types=1);
 
 namespace Hyde\Framework\Testing\Unit\Views;
 
-use Hyde\Framework\Helpers\Asset;
+use Hyde\Helpers\Asset;
 use Hyde\Hyde;
 use Hyde\Testing\TestCase;
 use Illuminate\Support\Facades\Blade;
